pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "matkit"
version = "0.2.0"
description = "A robust Python toolkit for materials informatics with advanced features"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "MatKit Contributors", email = "contact@matkit.org"}
]
keywords = ["materials", "informatics", "machine learning", "chemistry", "physics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Information Analysis"
]

dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "pymatgen>=2023.1.1",
    "mp-api>=0.30.0",
    "ase>=3.22.0",
    "torch>=2.0.0",
    "dscribe>=2.0.0",
    "plotly>=5.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "tenacity>=8.0.0",
    "rich>=13.0.0",
    "joblib>=1.2.0",
    "networkx>=3.0",
    "scipy>=1.9.0",
    "monty>=2023.1.1",
    "tqdm>=4.65.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.2.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.20.0",
    "black>=23.0.0",
    "ruff>=0.0.250",
    "mypy>=1.0.0",
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "nbsphinx>=0.8.0",
    "ipython>=8.0.0",
    "jupyter>=1.0.0",
    "pre-commit>=3.0.0"
]

ml = [
    "optuna>=3.0.0",
    "catboost>=1.1.0",
    "xgboost>=1.7.0",
    "lightgbm>=3.3.0",
    "shap>=0.41.0"
]

viz = [
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
    "crystal-toolkit>=2023.1.1"
]

[project.urls]
"Homepage" = "https://github.com/your-org/matkit"
"Documentation" = "https://matkit.readthedocs.io"
"Repository" = "https://github.com/your-org/matkit"
"Bug Tracker" = "https://github.com/your-org/matkit/issues"

[tool.setuptools]
packages = ["matkit"]

[tool.setuptools.package-data]
matkit = ["py.typed", "data/*.json"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "UP", "S", "B", "A", "C4", "RET", "SIM"]
ignore = ["E501", "S101"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true